requirements.txt
python-dotenv
requests
argparse

# brew install imagemagick